# contracts/analysis_service/api.yaml
openapi: 3.0.0
info:
  title: Analysis service API
  version: 0.1.0

paths:
  /zones/{zoneID}/analise:
    get:
      summary: Get zone analyses info
      parameters:
        - name: zoneID
          in: path
          required: true
          description: The id of the zone
          schema:
            type: string
      responses:
        '200':
          description: Expected zone analyse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneDetails'
        '400':
          description: Zone id is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /zones/{zoneID}/like/{userID}:
    post:
      summary: Add zone to list of favorites
      parameters:
        - name: zoneID
          in: path
          required: true
          description: The id of the zone
          schema:
            type: string
        - name: userID
          in: path
          required: true
          description: The id of the user
          schema:
            type: string
      responses:
        '201':
          description: Zone liked successfully
        '400':
          description: Zone id is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete zone from list of favorites
      parameters:
        - name: zoneID
          in: path
          required: true
          description: The id of the zone
          schema:
            type: string
        - name: userID
          in: path
          required: true
          description: The id of the user
          schema:
            type: string
      responses:
        '200':
          description: Zone disliked successfully
        '400':
          description: Zone id is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/zones:
    get:
      summary: Get favorite zones
      responses:
        '200':
          description: List of liked zones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zones'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/auth/{userID}:
    get:
      summary: Auth user
      parameters:
        - name: userID
          in: path
          required: true
          description: The id of the user
          schema:
            type: string
      responses:
        '200':
          description: User role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/create/{userID}:
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - userName
              properties:
                role:
                  type: string
                userName:
                  type: string
      parameters:
        - name: userID
          in: path
          required: true
          description: The id of the user
          schema:
            type: string
      responses:
        '201':
          description: User created
        '500':
          description: Failed to create user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    ZoneDetails:
      type: object
      required:
        - id
        - propertyType
        - permittedUsage
      properties:
        id:
          type: string
        propertyType:
          $ref: '#/components/schemas/PropertyType'
        rightType:
          type: string
        permittedUsage:
          type: string
        address:
          type: string
        square:
          type: integer
    Zone:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    Zones:
      type: array
      items:
        $ref: '#/components/schemas/Zone'
    Role:
      type: string
      enum:
        - Admin
        - Analyser
        - UndefinedRole
    PropertyType:
      type: string
      enum:
        - Undefined
        - Private
        - Municipal
        - State
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string