# generated/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// Defines values for PropertyType.
const (
	Municipal PropertyType = "Municipal"
	Private   PropertyType = "Private"
	State     PropertyType = "State"
	Undefined PropertyType = "Undefined"
)

// Defines values for Role.
const (
	Admin         Role = "Admin"
	Analyser      Role = "Analyser"
	UndefinedRole Role = "UndefinedRole"
)

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// PropertyType defines model for PropertyType.
type PropertyType string

// Role defines model for Role.
type Role string

// Zone defines model for Zone.
type Zone struct {
	Id string `json:"id"`
}

// ZoneDetails defines model for ZoneDetails.
type ZoneDetails struct {
	Address        *string      `json:"address,omitempty"`
	Id             string       `json:"id"`
	PermittedUsage string       `json:"permittedUsage"`
	PropertyType   PropertyType `json:"propertyType"`
	RightType      *string      `json:"rightType,omitempty"`
	Square         *int         `json:"square,omitempty"`
}

// Zones defines model for Zones.
type Zones = []Zone

// PostUserCreateUserIDJSONBody defines parameters for PostUserCreateUserID.
type PostUserCreateUserIDJSONBody struct {
	Role     string `json:"role"`
	UserName string `json:"userName"`
}

// PostUserCreateUserIDJSONRequestBody defines body for PostUserCreateUserID for application/json ContentType.
type PostUserCreateUserIDJSONRequestBody PostUserCreateUserIDJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Auth user
	// (GET /user/auth/{userID})
	GetUserAuthUserID(ctx echo.Context, userID string) error
	// Create user
	// (POST /user/create/{userID})
	PostUserCreateUserID(ctx echo.Context, userID string) error
	// Get favorite zones
	// (GET /user/zones)
	GetUserZones(ctx echo.Context) error
	// Get zone analyses info
	// (GET /zones/{zoneID}/analise)
	GetZonesZoneIDAnalise(ctx echo.Context, zoneID string) error
	// Delete zone from list of favorites
	// (DELETE /zones/{zoneID}/like/{userID})
	DeleteZonesZoneIDLikeUserID(ctx echo.Context, zoneID string, userID string) error
	// Add zone to list of favorites
	// (POST /zones/{zoneID}/like/{userID})
	PostZonesZoneIDLikeUserID(ctx echo.Context, zoneID string, userID string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetUserAuthUserID converts echo context to params.
func (w *ServerInterfaceWrapper) GetUserAuthUserID(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "userID" -------------
	var userID string

	err = runtime.BindStyledParameterWithOptions("simple", "userID", ctx.Param("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetUserAuthUserID(ctx, userID)
	return err
}

// PostUserCreateUserID converts echo context to params.
func (w *ServerInterfaceWrapper) PostUserCreateUserID(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "userID" -------------
	var userID string

	err = runtime.BindStyledParameterWithOptions("simple", "userID", ctx.Param("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostUserCreateUserID(ctx, userID)
	return err
}

// GetUserZones converts echo context to params.
func (w *ServerInterfaceWrapper) GetUserZones(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetUserZones(ctx)
	return err
}

// GetZonesZoneIDAnalise converts echo context to params.
func (w *ServerInterfaceWrapper) GetZonesZoneIDAnalise(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "zoneID" -------------
	var zoneID string

	err = runtime.BindStyledParameterWithOptions("simple", "zoneID", ctx.Param("zoneID"), &zoneID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter zoneID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetZonesZoneIDAnalise(ctx, zoneID)
	return err
}

// DeleteZonesZoneIDLikeUserID converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteZonesZoneIDLikeUserID(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "zoneID" -------------
	var zoneID string

	err = runtime.BindStyledParameterWithOptions("simple", "zoneID", ctx.Param("zoneID"), &zoneID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter zoneID: %s", err))
	}

	// ------------- Path parameter "userID" -------------
	var userID string

	err = runtime.BindStyledParameterWithOptions("simple", "userID", ctx.Param("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteZonesZoneIDLikeUserID(ctx, zoneID, userID)
	return err
}

// PostZonesZoneIDLikeUserID converts echo context to params.
func (w *ServerInterfaceWrapper) PostZonesZoneIDLikeUserID(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "zoneID" -------------
	var zoneID string

	err = runtime.BindStyledParameterWithOptions("simple", "zoneID", ctx.Param("zoneID"), &zoneID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter zoneID: %s", err))
	}

	// ------------- Path parameter "userID" -------------
	var userID string

	err = runtime.BindStyledParameterWithOptions("simple", "userID", ctx.Param("userID"), &userID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter userID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostZonesZoneIDLikeUserID(ctx, zoneID, userID)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/user/auth/:userID", wrapper.GetUserAuthUserID)
	router.POST(baseURL+"/user/create/:userID", wrapper.PostUserCreateUserID)
	router.GET(baseURL+"/user/zones", wrapper.GetUserZones)
	router.GET(baseURL+"/zones/:zoneID/analise", wrapper.GetZonesZoneIDAnalise)
	router.DELETE(baseURL+"/zones/:zoneID/like/:userID", wrapper.DeleteZonesZoneIDLikeUserID)
	router.POST(baseURL+"/zones/:zoneID/like/:userID", wrapper.PostZonesZoneIDLikeUserID)

}